requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
black==23.3.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
fsspec==2023.5.0
gast==0.4.0
gensim==4.3.1
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth-stubs==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpc-stubs==1.53.0.2
grpcio==1.54.2
h5py==3.8.0
httplib2==0.22.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
jax==0.4.10
jedi==0.18.2
joblib==1.2.0
jupyter-client==8.2.0
jupyter-core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mypy==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.1.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
types-requests==2.31.0.0
types-urllib3==1.26.25.13
typing-extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
